# user-service/docs/openapi.yaml
components:
    schemas:
...
        CreateRoleRequest:
            properties:
                active:
                    example: true
                    type: boolean
                description:
                    example: 管理后台角色
                    maxLength: 512
                    type: string
                name:
                    example: 管理员
                    maxLength: 128
                    minLength: 1
                    type: string
                system:
                    example: false
                    type: boolean
            required:
                - name
            type: object
...
        Envelope:
            properties:
                code:
                    $ref: '#/components/schemas/Code'
                data: {}
                errors: {}
                message:
                    type: string
                success:
                    type: boolean
            type: object
...
        Pagination:
            properties:
                has_next:
                    example: true
                    type: boolean
                next_cursor:
                    example: 0190c8d2-8d8a-7a01-9f43-0f91fb4e2b7c
                    type: string
                page_size:
                    example: 50
                    type: integer
            type: object
...
        RoleListResponseDoc:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/RoleResponse'
                    type: array
                pagination:
                    $ref: '#/components/schemas/Pagination'
            type: object
        RoleResponse:
            properties:
                active:
                    example: true
                    type: boolean
                created_at:
                    example: 1.7802888e+12
                    type: integer
                description:
                    example: 管理后台角色
                    type: string
                name:
                    example: 管理员
                    type: string
                role_id:
                    example: 018f6f3e-7c4d-7b2a-9f8a-4f6b1b2c3d4e
                    type: string
                system:
                    example: false
                    type: boolean
                updated_at:
                    example: 1.7802888e+12
                    type: integer
            type: object
...
        SetRoleStatusRequest:
            properties:
                active:
                    example: true
                    type: boolean
            type: object
...
        UpdateRoleRequest:
            properties:
                active:
                    example: true
                    type: boolean
                description:
                    example: 管理后台角色
                    maxLength: 512
                    type: string
                name:
                    example: 管理员
                    maxLength: 128
                    minLength: 1
                    type: string
            required:
                - name
            type: object
...
    securitySchemes:
        BearerAuth:
            bearerFormat: JWT
            description: 输入 Bearer token，格式为：Bearer <token>
            scheme: bearer
            type: http
info:
    contact: {}
    description: AegisCore 用户服务 API 文档，覆盖认证会话、用户资料、角色管理、权限目录、RBAC 授权保护的业务接口和服务健康检查。
    title: AegisCore User Services API
    version: 1.0.0
openapi: 3.0.3
paths:
...
    /roles:
        get:
            description: 查询角色列表。业务接口由 JWT 和 RBAC 保护，Casbin 使用 role_id 作为策略主体，不依赖 roles.code。
            parameters:
                - description: 分页游标
                  in: query
                  name: cursor
                  schema:
                    type: string
                - description: 每页数量
                  in: query
                  name: page_size
                  schema:
                    type: integer
                - description: 是否启用
                  in: query
                  name: active
                  schema:
                    type: boolean
                - description: 是否系统角色
                  in: query
                  name: system
                  schema:
                    type: boolean
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/Envelope'
                                    - properties:
                                        data:
                                            $ref: '#/components/schemas/RoleListResponseDoc'
                                      type: object
                    description: 查询成功
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 查询参数错误
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 未认证或 token 无效
                "403":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 无访问权限
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 服务器内部错误
            security:
                - BearerAuth: []
            summary: 分页查询角色
            tags:
                - 角色
        post:
            description: 创建角色。业务接口由 JWT 和 RBAC 保护。
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateRoleRequest'
                description: 创建角色请求
                required: true
                x-originalParamName: request
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/Envelope'
                                    - properties:
                                        data:
                                            $ref: '#/components/schemas/RoleResponse'
                                      type: object
                    description: 创建成功
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 请求体错误或参数校验失败
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 未认证或 token 无效
                "403":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 无访问权限
                "409":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 角色已存在
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 服务器内部错误
            security:
                - BearerAuth: []
            summary: 创建角色
            tags:
                - 角色
    /roles/{role_id}:
        get:
            description: 通过角色 ID 查询角色详情。业务接口由 JWT 和 RBAC 保护。
            parameters:
                - description: 角色ID
                  in: path
                  name: role_id
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/Envelope'
                                    - properties:
                                        data:
                                            $ref: '#/components/schemas/RoleResponse'
                                      type: object
                    description: 查询成功
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 角色 ID 参数错误
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 未认证或 token 无效
                "403":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 无访问权限
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 角色不存在
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 服务器内部错误
            security:
                - BearerAuth: []
            summary: 查询角色详情
            tags:
                - 角色
        patch:
            description: 更新角色元数据。系统角色受保护字段不可破坏性修改。
            parameters:
                - description: 角色ID
                  in: path
                  name: role_id
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateRoleRequest'
                description: 更新角色请求
                required: true
                x-originalParamName: request
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/Envelope'
                                    - properties:
                                        data:
                                            $ref: '#/components/schemas/RoleResponse'
                                      type: object
                    description: 更新成功
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 请求错误或系统角色受保护
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 未认证或 token 无效
                "403":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 无访问权限
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 角色不存在
                "409":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 角色已存在
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 服务器内部错误
            security:
                - BearerAuth: []
            summary: 更新角色
            tags:
                - 角色
...
    /roles/{role_id}/permissions/{permission_id}:
        delete:
            description: 删除角色权限绑定，解绑后相关 RBAC 授权应被拒绝。
            parameters:
                - description: 角色ID
                  in: path
                  name: role_id
                  required: true
                  schema:
                    type: string
                - description: 权限ID
                  in: path
                  name: permission_id
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/Envelope'
                                    - properties:
                                        data:
                                            items:
                                                $ref: '#/components/schemas/PermissionResponse'
                                            type: array
                                      type: object
                    description: 解绑成功
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 请求错误
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 未认证或 token 无效
                "403":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 无访问权限
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 绑定不存在
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 服务器内部错误
            security:
                - BearerAuth: []
            summary: 解绑角色权限
            tags:
                - 角色
    /roles/{role_id}/status:
        patch:
            description: 启用或停用角色。停用角色后相关 RBAC 授权应被拒绝。
            parameters:
                - description: 角色ID
                  in: path
                  name: role_id
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/SetRoleStatusRequest'
                description: 角色状态请求
                required: true
                x-originalParamName: request
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/Envelope'
                                    - properties:
                                        data:
                                            $ref: '#/components/schemas/RoleResponse'
                                      type: object
                    description: 操作成功
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 请求错误或系统角色受保护
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 未认证或 token 无效
                "403":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 无访问权限
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 角色不存在
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 服务器内部错误
            security:
                - BearerAuth: []
            summary: 启停角色
            tags:
                - 角色
...
    /users/{user_id}/roles/{role_id}:
        delete:
            description: 删除用户角色绑定，解绑后相关 RBAC 授权应被拒绝。
            parameters:
                - description: 用户ID
                  in: path
                  name: user_id
                  required: true
                  schema:
                    type: string
                - description: 角色ID
                  in: path
                  name: role_id
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/Envelope'
                                    - properties:
                                        data:
                                            items:
                                                $ref: '#/components/schemas/RoleResponse'
                                            type: array
                                      type: object
                    description: 解绑成功
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 请求错误
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 未认证或 token 无效
                "403":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 无访问权限
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 绑定不存在
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Envelope'
                    description: 服务器内部错误
            security:
                - BearerAuth: []
            summary: 解绑用户角色
            tags:
                - 角色
servers:
    - url: /api/v1